pkg/dhclient: refactor lease configuration.
- Move lease configure4&6 to their respective file.
- Update v6 configuration call to be consistent with v4.
